### 2023-08-04 #5736 @editorialbot
[PRE REVIEW]: Seaduck: A python package for Eulerian and Lagrangian interpolation on ocean datasets
**Submitting author:** @MaceKuailv (<a href="http://orcid.org/0009-0000-7319-9469">Wenrui Jiang</a>)
**Repository:** https://github.com/MaceKuailv/seaduck
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
