### 2023-08-04 #5736 @kthyng
Initial pre-review issue: https://github.com/openjournals/joss-reviews/issues/5726#event-9990663341
