+++ .pre-commit-config.yaml
@@ -25,7 +25,7 @@ repos:
   - id: blackdoc
     additional_dependencies: [black==22.3.0]
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.0.281
+  rev: v0.0.284
   hooks:
   - id: ruff
     args: [--fix]
