### 2023-08-17 7aacda6
fix a bug in ind_tend_uv
