### 2023-08-17 6181c97
fix a bug in ind_tend_uv
