+++ .github/workflows/draft-pdf.yml
@@ -5,7 +5,7 @@ jobs:
     name: Paper Draft
     steps:
     - name: Checkout
-      uses: actions/checkout@v3
+      uses: actions/checkout@v4
     - name: Build draft PDF
       uses: openjournals/openjournals-draft-action@master
       with:
