### 2023-09-15 3262f24
stringent prefetch update to fix issue 85
