### 2023-10-09 #5736 @kthyng
Hi @pnsaevik and @syedhamidali! Are you interested in reviewing this JOSS submission? I would ask that you start your review in the next 3 weeks with the intention of having it finished within 4-6 weeks, given that the process is often iterative with the submitting author(s). You can get more information about reviewing for JOSS here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#

Thanks for your consideration!
