### 2023-10-16 #5736 @pnsaevik
Hi @kthyng, I accept the invitation. This is my first time reviewing for JOSS, but I'm familiar with the process through my own submission previously.
