### 2023-10-20 #5736 @syedhamidali
Hi @kthyng, I accept the invitation with enthusiasm. This is my first time reviewing for JOSS, and while I'm excited to begin, my semester is quite busy. Please understand if my responses take a bit longer. I'll ensure a thorough review. Thanks for your patience.
