### 2023-10-20 #5736 @kthyng
Excellent @pnsaevik and @syedhamidali! @syedhamidali I find it's most effective when reviewers are able to be responsive to questions from authors and make even small steps of progress to move things along since often questions or interactions come up in the process and you'll need to wait for something anyway.

I'll start the review now.
