### 2023-10-20 #5967 @editorialbot
[REVIEW]: Seaduck: A python package for Eulerian and Lagrangian interpolation on ocean datasets
**Submitting author:** @MaceKuailv (<a href="http://orcid.org/0009-0000-7319-9469">Wenrui Jiang</a>)
**Repository:** https://github.com/MaceKuailv/seaduck
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
