### 2023-10-26 #5967 @MaceKuailv
We were experiencing a small issue with zenodo and pooch (the way our unit tests download the datasets) a few days ago. Now, the folks working on pooch has sort that out, so everything should be working now.
