### 2023-10-30 MaceKuailv/seaduck#90 @pnsaevik
Incomplete requirements specification
The list of required and optional dependencies is outdated. The listed optional dependencies are already included in the pip required dependencies list, except matplotlib which is not used in the package itself as far as I can see. But the "required dependencies" list can be removed from the documentation I think, as long as the pip installation process takes care of it. If extra installation steps are required (on Windows for instance), these should be included here.

The optional dependency list should include packages that are required to run the tutorials (zarr, pooch etc.).

This issue is a part of a JOSS review at openjournals/joss-reviews#5967
