### 2023-10-30 MaceKuailv/seaduck#90 @pnsaevik
Also, when listing optional dependencies, it should be clearly defined what the dependency is used for. This way, users can know whether they need to install the dependency or not.
