### 2023-10-30 MaceKuailv/seaduck#91 @pnsaevik
Docstring errors are not captured by tests
Docstring errors are not captured by tests. There is a current docstring formatting error in the function `seaduck.ocedata.OceData.check_readiness` which is visible in the build log of `jupyter-book build`, but does not generate a test fail.

This issue is a part of a JOSS review at https://github.com/openjournals/joss-reviews/issues/5967
