### 2023-10-30 MaceKuailv/seaduck#92 @pnsaevik
Tutorial errors are not captured by tests
I intentionally introduced an error to one of the tutorial files. The error shows up as a warning in the jupyter-book build log, but does not generate a nonzero exit value and is therefore not captured by the CI testing framework. It is important that tutorials are regularly tested, so that any changes to the API that breaks the examples are immediately reported to developers.

This issue is a part of a JOSS review at https://github.com/openjournals/joss-reviews/issues/5967
