### 2023-10-30 MaceKuailv/seaduck#93 @pnsaevik
Missing instructions on issue reports
The documentation contains information on how to contribute with source code, but not on how to report issues and how so seek support (as required by JOSS).

This is trivial to fix, for instance: Add a new heading under "Contributing" named "Report issues" or similar, mentioning that the GitHub issue tracker should be used. Also mention that limited support can be given through the same channel, or state explicitly that support on how to use the software will not be given (it's free software, after all).

This issue is a part of a JOSS review at https://github.com/openjournals/joss-reviews/issues/5967
