### 2023-10-30 MaceKuailv/seaduck#94 @pnsaevik
Development status
In my opinion, a JOSS submission should be marked as `Development Status :: 5 - Production/Stable` on PyPI, and it's version number should not start with 0 (see also, e.g., https://martin-thoma.com/software-development-stages/).

This package is already quite mature, so there is no reason it should not be marked as "Stable", even with the issues I previously raised on GitHub. With one exception: It should be clear to the user which part of the API is considered stable, and which is open to change. At the very least, examples in the tutorials should have "frozen syntax" and should not be changed except at infrequent major version upgrades.

Before bumping up the release to version 1.0.0, I suggest you do the following:

1. Fix minor issues already raised
2. Clearly state which part of the API is fixed, either summarized on the API Reference front page or individually for each exposed function.
3. If you are unhappy with any part of the public API, now is the time to fix it. Do an internal review among developers and internal users:
3.1. Are the variable/class/function/parameter names intuitive?
3.2. Is the naming scheme consistent?
3.3. Are we happy with the module/class organization? Should something be moved?
3.4. Remember that parameter names, their ordering, and possible default values are part of the public API. Go through each function and make changes if necessary, before declaring the function as a part of the public "frozen" API.

This issue is a part of a JOSS review at https://github.com/openjournals/joss-reviews/issues/5967
