### 2023-10-30 MaceKuailv/seaduck#92 @MaceKuailv
Yes! I wanted to have that a while ago, but didn't know how to achieve it. I will revisit it today.
