### 2023-10-30 c94b9ea
specify path for jupyter log
