### 2023-10-30 1babb07
specify path for jupyter log
