### 2023-10-30 f3f62e0
found a way to raise all warnings in jupyter book as exit code 1
