### 2023-10-30 a44c8cd
found a way to raise all warnings in jupyter book as exit code 1
