### 2023-10-30 5c373b9
fix check_readiness docstring
