+++ docs/_config.yml
@@ -1,4 +1,4 @@
-# Book settings
+/# Book settings
 # Learn more at https://jupyterbook.org/customize/config.html
 
 title: My sample book
