+++ docs/_config.yml
@@ -1,8 +1,8 @@
-/# Book settings
+# Book settings
 # Learn more at https://jupyterbook.org/customize/config.html
 
-title: My sample book
-author: The Jupyter Book Community
+title: Seaduck Tutorial
+author: Wenrui Jiang
 logo: logo.png
 
 # Force re-execution of notebooks on each build.
@@ -11,8 +11,6 @@ execute:
   timeout: 300
   execute_notebooks: 'auto'
   stderr_output: "remove"
-  warn_if:
-  - error
   exclude_patterns:
   - '*.md'
 
@@ -33,10 +31,6 @@ sphinx:
     #     "relbarbgcolor": 'Orange',
     #     "codebgcolor": 'Coral'
     # }
-
-# Add a bibtex file so that we can create citations
-bibtex_bibfiles:
-- references.bib
 
 # Information about where the book exists on the web
 repository:
