### 2023-10-30 MaceKuailv/seaduck#94 @MaceKuailv
Thank you! I was not aware of the convention of releases. This is very helpful. I think I will bump it up to 1.0.0 when both of you are happy with it.

I don't think there is going to be a lot of changes to the syntax or keyword arguments on functions and classes used in the tutorial examples. Before submitting this paper, we spend a lot of time discussing pretty much exactly the third item and already made a lot of changes. For example, PR #58.

I am not entirely sure what you mean by labeling the part of API that is fixed. Do you have an example on how other people did that? Thanks!
