### 2023-10-30 0f63218
fix a warning from sphinx theme
