+++ docs/_config.yml
@@ -43,3 +43,6 @@ repository:
 html:
   use_issues_button: true
   use_repository_button: true
+
+html_theme_options:
+  navigation_with_keys: false
