### 2023-10-30 57405ca
try another place to put the navigation-key thing
