+++ docs/how_to_install.md
@@ -14,15 +14,16 @@ conda install -c conda-forge seaduck
 
 or, alternatively (for latest code development), see [a guide for new contibutors](guide_for_developer.md).
 
+## Optional dependencies
+
+- numba
+  - Compile and make everything faster!
+- matplotlib
+  - Plot the shape of interpolation kernel.
+
 ## Required dependencies
 
 - python
 - numpy
 - xarray
 - dask
-
-## Optional dependencies
-
-- pandas
-- numba
-- matplotlib
