### 2023-10-30 5aaa772
conda install is available
