### 2023-10-30 MaceKuailv/seaduck#90 @MaceKuailv
Thanks for the suggestion. I have made some changes on that front.

`dask` became a required dependency pretty recently, so it is moved to where . We also didn't realize that `pandas` is a dependency of xarray, so now it is removed.

We don't have a lot of required dependencies, so instead of removing them I put them after the optional ones. Like you mentioned, they are not very important to users installing packages. However, environments and dependencies can get very nasty sometimes. I think it does not hurt to put some clue in the documentation.

I have tried installation on my windows laptop (~2 month ago) with anaconda. It does not require additional steps.
