### 2023-10-30 MaceKuailv/seaduck#90 @pnsaevik
As long as your package uses `pandas` explicitly, it makes sense to list it as a dependency (at least in pyproject.toml). If `xarray` becomes independent from `pandas` in the future, your package will break if `pandas` is not a requirement. (Okay, this is likely not gonna happen. But my point is that an explicit dependency is not necessarily a bad thing. Treat `pandas` like you would treat `numpy`, which is also required by most packages).

But don't forget the missing optional dependencies. As it stands, users cannot run the tutorials unless installing the extra packages. It should be clearly stated which packages are required to run the examples.
