+++ docs/contribute/report_iss.md
@@ -1,0 +1,16 @@
+# Report Issues on GitHub
+
+Aside from contributing to code and documentation, another good way to help the package is to report issues. An Issue could be about a feature request, a bug report, or any of your cool suggestions.
+
+We use GitHub to keep track of the issues. To see a list of all past issues, simply go to [this page](https://github.com/MaceKuailv/seaduck/issues). 
+
+Before you report a new issue, make sure you follow these two simple steps:
+
+1. Read through relevant sections in the documentation; 
+2. Use the search bar to see if the issue has already been reported.
+
+If both didn't help, you can go ahead and click the big green button that says "New Issue". Writing the issue should be very straightforward. Make sure that the message is clear to other people reading this issue. One of the best ways of doing that is to give a [minimum reproducible example](https://en.wikipedia.org/wiki/Minimal_reproducible_example). 
+
+Please remember that the purpose of this GitHub issues channel is to improve the package and that the developers of `seaduck` have their day-time jobs. We are not going to provide direct instruction on how to use this package or python. 
+
+That said, never feel intimidated to ask questions that you think is dumb. By reporting an issue, you also help others who are in the same situation. If you spend enough effort working on the problem and still cannot figure it out, there must be something confusing or wrong about the code or the documentation. We will really appreciate it if you point that out. 
