### 2023-10-30 d999556
improve optional dependency, after clarified
