+++ docs/how_to_install.md
@@ -20,6 +20,12 @@ or, alternatively (for latest code development), see [a guide for new contibutor
   - Compile and make everything faster!
 - matplotlib
   - Plot the shape of interpolation kernel.
+- pandas
+  - Show variable aliasing in an intuitive way.
+
+- pooch, zarr, cartopy
+  - Needed for running the tutorial examples.
+
 
 ## Required dependencies
 
