### 2023-10-30 MaceKuailv/seaduck#90 @MaceKuailv
Yes, that makes sense. I have just re-included pandas and added the packages needed for running the tutorial examples.

Part of the tutorials are ran on Sciserver, which is a publicly-accessible compute platform with readily available docker environments. There are a few packages like `tqdm` in the CI environment file that is never used in any of the non-Sciserver examples. They are also already installed on Sciserver. I have excluded them in the optional dependencies, because I think the users don't need to know about that.
