### 2023-10-30 MaceKuailv/seaduck#94 @MaceKuailv
OK, I think that's very doable and useful. I will make it clear in the table of content. I will also look into the public files and identify the non-public functions and underscore them.
