+++ seaduck/topology.py
@@ -29,7 +29,7 @@ DIRECTIONS = np.array([np.pi / 2, -np.pi / 2, np.pi, 0])
 
 
 @compileable
-def llc_mutual_direction(face, neighbor_face, transitive=False):
+def _llc_mutual_direction(face, neighbor_face, transitive=False):
     """Find the relative orientation of two faces.
 
     The compileable version of mutual direction for llc grid.
@@ -77,7 +77,7 @@ def llc_mutual_direction(face, neighbor_face, transitive=False):
 
 
 @compileable
-def llc_get_the_other_edge(face, edge):
+def _llc_get_the_other_edge(face, edge):
     """See what is adjacent to the face by this edge.
 
     The compileable version of get_the_other_edge for llc grid.
@@ -94,7 +94,7 @@ def llc_get_the_other_edge(face, edge):
 
 
 @compileable
-def box_ind_tend(ind, tend, iymax, ixmax):
+def _box_ind_tend(ind, tend, iymax, ixmax):
     """Move an index in a direction.
 
     The compileable version of ind_tend for regional (box) grid.
@@ -118,7 +118,7 @@ def box_ind_tend(ind, tend, iymax, ixmax):
 
 
 @compileable
-def x_per_ind_tend(ind, tend, iymax, ixmax):
+def _x_per_ind_tend(ind, tend, iymax, ixmax):
     """Move an index in a direction.
 
     The compileable version of ind_tend for zonally periodic (x-per) grid.
@@ -143,7 +143,7 @@ def x_per_ind_tend(ind, tend, iymax, ixmax):
 
 
 @compileable
-def llc_ind_tend(ind, tendency, iymax, ixmax):
+def _llc_ind_tend(ind, tendency, iymax, ixmax):
     """Move an index in a direction.
 
     The compileable version of ind_tend for llc grid.
@@ -156,7 +156,7 @@ def llc_ind_tend(ind, tendency, iymax, ixmax):
         if ix != ixmax:
             ix += 1
         else:
-            neighbor_face, new_edge = llc_get_the_other_edge(face, 3)
+            neighbor_face, new_edge = _llc_get_the_other_edge(face, 3)
             if new_edge == 1:
                 face, iy, ix = [neighbor_face, 0, ixmax - iy]
             elif new_edge == 0:
@@ -169,7 +169,7 @@ def llc_ind_tend(ind, tendency, iymax, ixmax):
         if ix != 0:
             ix -= 1
         else:
-            neighbor_face, new_edge = llc_get_the_other_edge(face, 2)
+            neighbor_face, new_edge = _llc_get_the_other_edge(face, 2)
             if new_edge == 1:
                 face, iy, ix = [neighbor_face, 0, iy]
             elif new_edge == 0:
@@ -182,7 +182,7 @@ def llc_ind_tend(ind, tendency, iymax, ixmax):
         if iy != iymax:
             iy += 1
         else:
-            neighbor_face, new_edge = llc_get_the_other_edge(face, 0)
+            neighbor_face, new_edge = _llc_get_the_other_edge(face, 0)
             if new_edge == 1:
                 face, iy, ix = [neighbor_face, 0, ix]
             elif new_edge == 0:
@@ -195,7 +195,7 @@ def llc_ind_tend(ind, tendency, iymax, ixmax):
         if iy != 0:
             iy -= 1
         else:
-            neighbor_face, new_edge = llc_get_the_other_edge(face, 1)
+            neighbor_face, new_edge = _llc_get_the_other_edge(face, 1)
             if new_edge == 1:
                 face, iy, ix = [neighbor_face, 0, ixmax - ix]
             elif new_edge == 0:
[…]
