+++ docs/api_reference.md
@@ -1,25 +1,0 @@
-# API Reference
-
-## Dataset-related object
-
-[OceData](api_reference/apiref_OceData.rst)
-[topology](api_reference/apiref_topology.rst)
-
-## Point/particle object
-
-[eulerian](api_reference/apiref_eulerian.rst)
-[lagrangian](api_reference/apiref_lagrangian.rst)
-
-## High-level function
-
-[OceInterp](api_reference/apiref_OceInterp.rst)
-
-## Kernel object
-
-[Kernal_and_Weight](api_reference/apiref_kernelNweight.rst)
-
-## Utils
-
-[smart_read](api_reference/apiref_smartread.rst)
-[get_masks](api_reference/apiref_getmasks.rst)
-[utils](api_reference/apiref_utils.rst)
