+++ docs/internal_api_reference.md
@@ -1,0 +1,10 @@
+# Internal API reference
+
+## utility functions
+
+[smart_read](api_reference/apiref_smartread.rst)
+[get_masks](api_reference/apiref_getmasks.rst)
+[utils](api_reference/apiref_utils.rst)
+
+## Relative coordinate handling
+[RelCoord](api_reference/apiref_RelCoord.rst)
