+++ docs/get_started.md
@@ -2,7 +2,7 @@
 
 For a extremely quick start, read this [one minute guide](./one_min_guide.md). Or, learn about how the objects and functions are related to each other [here](./network_of_object.md).
 
-If you have questions about any function, you can always go to [API references](./api_reference.md). Or you could checkout the following examples. More example can be found [here](ideal_test.md) and [here](ocean_example.md)
+If you have questions about any function, you can always go to [API references](./public_api_reference.md). Or you could checkout the following examples. More example can be found [here](ideal_test.md) and [here](ocean_example.md)
 
 ::::{grid} 1 1 1 3
 :class-container: text-center
