### 2023-10-30 MaceKuailv/seaduck#94 @MaceKuailv
The final changes I made in PR #95 implemented that. You can take a look after the website finishes building.
