### 2023-10-30 MaceKuailv/seaduck#90 @MaceKuailv
Resolved with PR #95
