### 2023-10-30 MaceKuailv/seaduck#91 @MaceKuailv
Resolved with PR #95
