### 2023-10-30 MaceKuailv/seaduck#92 @MaceKuailv
Resolved with PR #95
