### 2023-10-30 MaceKuailv/seaduck#93 @MaceKuailv
Resolved with PR #95
