+++ _sources/how_to_install.md
@@ -6,7 +6,7 @@ The package can be installed simply by
 pip install seaduck
 ```
 
-or, alternatively (not available yet)
+or, alternatively
 
 ```shell
 conda install -c conda-forge seaduck
@@ -14,15 +14,27 @@ conda install -c conda-forge seaduck
 
 or, alternatively (for latest code development), see [a guide for new contibutors](guide_for_developer.md).
 
+## Optional dependencies
+
+- numba
+
+  - Compile and make everything faster!
+
+- matplotlib
+
+  - Plot the shape of interpolation kernel.
+
+- pandas
+
+  - Show variable aliasing in an intuitive way.
+
+- pooch, zarr, cartopy
+
+  - Needed for running the tutorial examples.
+
 ## Required dependencies
 
 - python
 - numpy
 - xarray
-
-## Optional dependencies
-
-- pandas
-- numba
-- matplotlib
 - dask
