### 2023-10-31 MaceKuailv/seaduck#94 @pnsaevik
Great, the change to the documentation resolves the issue. There is still some potensial for polishing the documentation, but this is "further work".

If you want a small sample ROMS dataset for extending your package, take a look at https://github.com/pnsaevik/ladim_plugins/blob/master/ladim_plugins/chemicals/forcing.nc. It's an exercept from a fjord model in southern Norway.
