### 2023-10-31 MaceKuailv/seaduck#97 @pnsaevik
Broken links in the documentation
The link to the ETOPO dataset at https://macekuailv.github.io/seaduck/sciserver_notebooks/KangerFjord.html is broken. Ideally, there should be a test in place to warn developers about broken links. I tried the solution outlined in the [Jupyterbook documentation](https://jupyterbook.org/en/stable/advanced/html.html#html-link-check), but I could not get it to work. Maybe you have better luck?

This issue is a part of a JOSS review at https://github.com/openjournals/joss-reviews/issues/5967
