### 2023-10-31 MaceKuailv/seaduck#97 @MaceKuailv
Oh no, I thought those links are going to be pretty stable. It makes sense to have automatic checks. I will try to implement that.
