+++ Makefile
@@ -19,7 +19,7 @@ conda-env-update:
 	$(CONDA) env update $(CONDAFLAGS) -f environment.yml
 
 docs-build:
-	jupyter-book build docs/ -W --keep-going
+	jupyter-book build docs/ -W --keep-going --builder linkcheck
 
 joss:
 	pandoc paper/paper.md --bibliography paper/paper.bib -o paper/paper_local.pdf
