### 2023-10-31 c8dc2bc
fix broken link and introduce new one to see if it catch
