### 2023-10-31 194a6e1
remove the check and intentially wrong link, add instruction
