### 2023-10-31 8bacf2c
remove the check and intentially wrong link, add instruction
