+++ docs/intro.md
@@ -2,8 +2,6 @@
 
 Seaduck is a python package that helps you interpolate in an ocean dataset in both Eulerian and Lagrangian perspectives. Accessing ocean datasets with seaduck is as simple as throwing rubber ducks in the bathtub.
 
-[rogue link](www.jiangwenrui.com)
-
 "What have you cute duckling found down there? "
 
 "..."
