+++ docs/sciserver_notebooks/KangerFjord.md
@@ -30,8 +30,7 @@ import cmocean
 mpl.rcParams["figure.dpi"] = 300
 ```
 
-Here on [SciServer](https://sciserver.org/), we have an interesting dataset simulating the interaction between background circulation and Kangerdlugssuaq Fjord. More information can be found below, and see the paper by [^Fraser et al., 2018]
- [^Fraser et al., 2018]:10.1029/2018JC014435
+Here on [SciServer](https://sciserver.org/), we have an interesting dataset simulating the interaction between background circulation and Kangerdlugssuaq Fjord. More information can be found below, and see the paper by [Fraser et al., 2018](https://doi.org/10.1029/2018JC014435)
 
 ```{code-cell} ipython3
 fjord = ospy.open_oceandataset.from_catalog("KangerFjord")
