### 2023-10-31 MaceKuailv/seaduck#97 @MaceKuailv
I have added a optional manual check and fix the existing wrong links in PR #99.

However, I don't want to merge it just yet. This is because this is a pretty small fix and I would expect more from No. 99 :)
