### 2023-11-03 MaceKuailv/seaduck#100 @pnsaevik
Delete merged branches
This is a very minor issue, but best practice is to delete branches as soon as they are merged. There are 3 non-deleted, merged branches in the repo.

This issue is a part of a JOSS review at https://github.com/openjournals/joss-reviews/issues/5967
