### 2023-11-03 MaceKuailv/seaduck#101 @pnsaevik
Generate zenodo archive
JOSS prequires that the software is uploaded as a Zenodo archive as a part of the acceptance process. You may want to prepare this archive in anticipation of acceptance. You can wait until version 1.0.0 for uploading to zenodo, but you might as well do it now to check that everything works as expected (and re-upload the final version).

Zenodo looks for a [citation file](https://zenodo.org/records/5171937) in the repo. If the citation file is not present, you will have to add some metadata manually.

This issue is a part of a JOSS review at https://github.com/openjournals/joss-reviews/issues/5967
