### 2023-11-03 #5967 @pnsaevik
I'm done with my review and have no further issues. When @syedhamidali is done with his review, the package author will update the version number to 1.0.0 and upload the release to Zenodo, which is the only remaining issue on my part.

The package represents a novel contribution to the earth science community. It is based on sound numerical principles and good development practices.
