### 2023-11-05 MaceKuailv/seaduck#102 @syedhamidali
Conda installation link is either broken or not setup yet
I followed the installation instructions outlined in the [official documentation](https://macekuailv.github.io/seaduck/how_to_install.html) by attempting to install it via the `conda-forge` channel. However, I encountered an issue during the installation process.

This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/5967

Here is the issue:

```
conda install -c conda-forge seaduck
```
```
Collecting package metadata (current_repodata.json): done
Solving environment: unsuccessful initial attempt using frozen solve. Retrying with flexible solve.
Solving environment: unsuccessful initial attempt using frozen solve. Retrying with flexible solve.

PackagesNotFoundError: The following packages are not available from current channels:

  - seaduck

Current channels:

  - https://conda.anaconda.org/conda-forge/osx-64
  - https://repo.anaconda.com/pkgs/main/osx-64
  - https://repo.anaconda.com/pkgs/r/osx-64

To search for alternate channels that may provide the conda package you're
looking for, navigate to

    https://anaconda.org

and use the search bar at the top of the page.
```
