### 2023-11-06 0564612
change the documentation thumbnail link
