### 2023-11-07 8444de6
fix a ugly looking error message
