### 2023-11-07 MaceKuailv/seaduck#102 @syedhamidali
```zsh
syed44@MacBook-Pro-6 ~ % sw_vers

ProductName:		macOS
ProductVersion:		13.4
BuildVersion:		22F66

(base) syed44@MacBook-Pro-6 ~ % conda --version

conda 23.7.3
```

I am using macOS 13.4 on an M1 machine.
