### 2023-11-08 MaceKuailv/seaduck#102 @MaceKuailv
@syedhamidali Have you tried installing it again?
