### 2023-11-08 MaceKuailv/seaduck#102 @MaceKuailv
@syedhamidali Can you also try
```
conda search -c conda-forge seaduck
```
