### 2023-11-08 MaceKuailv/seaduck#102 @syedhamidali
@MaceKuailv, I attempted to create a new environment, but unfortunately, I encountered the same error. Here's what I tried:

```zsh
conda create -n seaduck-env
conda activate seaduck-env
conda install -c conda-forge seaduck
```

I also attempted the following approaches:

```zsh
conda create -n seaduck-env seaduck -c conda-forge
```
```zsh
conda create -n seaduck-env seaduck python=3.11 -c conda-forge
```
