### 2023-11-08 MaceKuailv/seaduck#102 @syedhamidali
> @syedhamidali Can you also try
>
> ```
> conda search -c conda-forge seaduck
> ```

```zsh
syed44@MacBook-Pro-6 ~ % conda search -c conda-forge seaduck
Loading channels: done
No match found for: seaduck. Search: *seaduck*

PackagesNotFoundError: The following packages are not available from current channels:

  - seaduck

Current channels:

  - https://conda.anaconda.org/conda-forge/osx-64
  - https://conda.anaconda.org/conda-forge/noarch
  - https://repo.anaconda.com/pkgs/main/osx-64
  - https://repo.anaconda.com/pkgs/main/noarch
  - https://repo.anaconda.com/pkgs/r/osx-64
  - https://repo.anaconda.com/pkgs/r/noarch

To search for alternate channels that may provide the conda package you're
looking for, navigate to

    https://anaconda.org

and use the search bar at the top of the page.
```
