### 2023-11-08 MaceKuailv/seaduck#102 @syedhamidali
However, I would like to mention that `mamba install seaduck` is working, but I am yet to test it.

```zsh
syed44@MacBook-Pro-6 ~ % mamba install seaduck

Looking for: ['seaduck']

pkgs/main/noarch                                              No change
pkgs/r/osx-arm64                                              No change
pkgs/main/osx-arm64                                  1.9MB @   5.4MB/s  0.4s
pkgs/r/noarch                                        2.0MB @   2.7MB/s  0.6s
conda-forge/noarch                                  12.5MB @  16.3MB/s  0.8s
conda-forge/osx-arm64                                7.3MB @   8.2MB/s  0.9s
Transaction

  Prefix: /Users/syed44/miniconda3/envs/bubblebath

  Updating specs:

   - seaduck

  Package                    Version  Build                 Channel           Size
────────────────────────────────────────────────────────────────────────────────────
  Install:
────────────────────────────────────────────────────────────────────────────────────

  + libexpat                   2.5.0  hb7217d7_1            conda-forge     Cached
  + python_abi                  3.12  4_cp312               conda-forge     Cached
  + libcxx                    16.0.6  h4653b0c_0            conda-forge     Cached
  + yaml                       0.2.5  h3422bc3_2            conda-forge     Cached
  + bzip2                      1.0.8  h93a5062_5            conda-forge      122kB
  + libffi                     3.4.2  h3422bc3_5            conda-forge     Cached
  + libzlib                   1.2.13  h53f4e23_5            conda-forge     Cached
  + xz                         5.2.6  h57fd34a_0            conda-forge     Cached
  + ncurses                      6.4  h463b476_2            conda-forge     Cached
  + ca-certificates        2023.7.22  hf0a4a13_0            conda-forge     Cached
  + llvm-openmp               17.0.4  hcd81f8e_0            conda-forge     Cached
  + tk                        8.6.13  h5083fa2_1            conda-forge        3MB
  + libsqlite                 3.44.0  h091b4b1_0            conda-forge     Cached
  + readline                     8.2  h92ec313_1            conda-forge     Cached
  + openssl                    3.1.4  h0d3ecfb_0            conda-forge     Cached
  + libgfortran5              13.2.0  hf226fd6_1            conda-forge     Cached
  + libgfortran                5.0.0  13_2_0_hd922786_1     conda-forge     Cached
  + libopenblas               0.3.24  openmp_hd76b1f2_0     conda-forge     Cached
  + libblas                    3.9.0  19_osxarm64_openblas  conda-forge     Cached
  + libcblas                   3.9.0  19_osxarm64_openblas  conda-forge     Cached
  + liblapack                  3.9.0  19_osxarm64_openblas  conda-forge     Cached
  + tzdata                     2023c  h71feb2d_0            conda-forge     Cached
  + python                    3.12.0  h47c9636_0_cpython    conda-forge     Cached
  + wheel                     0.41.3  pyhd8ed1ab_0          conda-forge     Cached
  + setuptools                68.2.2  pyhd8ed1ab_0          conda-forge     Cached
  + pip                       23.3.1  pyhd8ed1ab_0          conda-forge     Cached
  + six                       1.16.0  pyh6c4a22f_0          conda-forge     Cached
  + pytz                2023.3.post1  pyhd8ed1ab_0          conda-forge     Cached
  + python-tzdata             2023.3  pyhd8ed1ab_0          conda-forge     Cached
  + zipp                      3.17.0  pyhd8ed1ab_0          conda-forge     Cached
  + locket                     1.0.0  pyhd8ed1ab_0          conda-forge     Cached
  + toolz                     0.12.0  pyhd8ed1ab_0          conda-forge     Cached
  + packaging                   23.2  pyhd8ed1ab_0          conda-forge     Cached
  + fsspec                 2023.10.0  pyhca7485f_0          conda-forge     Cached
  + cloudpickle                3.0.0  pyhd8ed1ab_0          conda-forge     Cached
  + click                      8.1.7  unix_pyh707e725_0     conda-forge     Cached
  + python-dateutil            2.8.2  pyhd8ed1ab_0          conda-forge     Cached
  + importlib-metadata         6.8.0  pyha770c72_0          conda-forge     Cached
  + partd                      1.4.1  pyhd8ed1ab_0          conda-forge     Cached
  + importlib_metadata         6.8.0  hd8ed1ab_0            conda-forge     Cached
  + numpy                     1.26.0  py312h696b312_0       conda-forge     Cached
  + pyyaml                     6.0.1  py312h02f2b3b_1       conda-forge     Cached
  + pandas                     2.1.2  py312h9e53831_0       conda-forge     Cached
  + scipy                     1.11.3  py312h44883eb_1       conda-forge     Cached
  + dask-core              2023.10.1  pyhd8ed1ab_0          conda-forge     Cached
  + xarray                 2023.10.1  pyhd8ed1ab_0          conda-forge     Cached
  + seaduck                    0.1.3  pyhd8ed1ab_0          conda-forge       53kB

  Summary:

  Install: 47 packages

  Total download: 3MB

────────────────────────────────────────────────────────────────────────────────────

Confirm changes: [Y/n] y
tk                                                   3.1MB @   5.7MB/s  0.6s
bzip2                                              122.3kB @ 146.2kB/s  0.8s
seaduck                                             53.0kB @  54.2kB/s  1.0s

Downloading and Extracting Packages

Preparing transaction: done
Verifying transaction: done
Executing transaction: done
(bubblebath) syed44@MacBook-Pro-6 ~ % conda deactivate
(trop_dev) syed44@MacBook-Pro-6 ~ % conda remove --name bubblebath --all

Remove all packages in environment /Users/syed44/miniconda3/envs/bubblebath:

## Package Plan ##

  environment location: /Users/syed44/miniconda3/envs/bubblebath

The following packages will be REMOVED:

  bzip2-1.0.8-h93a5062_5
  ca-certificates-2023.7.22-hf0a4a13_0
  click-8.1.7-unix_pyh707e725_0
  cloudpickle-3.0.0-pyhd8ed1ab_0
  dask-core-2023.10.1-pyhd8ed1ab_0
  fsspec-2023.10.0-pyhca7485f_0
  importlib-metadata-6.8.0-pyha770c72_0
  importlib_metadata-6.8.0-hd8ed1ab_0
  libblas-3.9.0-19_osxarm64_openblas
  libcblas-3.9.0-19_osxarm64_openblas
  libcxx-16.0.6-h4653b0c_0
  libexpat-2.5.0-hb7217d7_1
  libffi-3.4.2-h3422bc3_5
  libgfortran-5.0.0-13_2_0_hd922786_1
  libgfortran5-13.2.0-hf226fd6_1
  liblapack-3.9.0-19_osxarm64_openblas
  libopenblas-0.3.24-openmp_hd76b1f2_0
  libsqlite-3.44.0-h091b4b1_0
  libzlib-1.2.13-h53f4e23_5
  llvm-openmp-17.0.4-hcd81f8e_0
  locket-1.0.0-pyhd8ed1ab_0
  ncurses-6.4-h463b476_2
  numpy-1.26.0-py312h696b312_0
  openssl-3.1.4-h0d3ecfb_0
  packaging-23.2-pyhd8ed1ab_0
  pandas-2.1.2-py312h9e53831_0
  partd-1.4.1-pyhd8ed1ab_0
  pip-23.3.1-pyhd8ed1ab_0
  python-3.12.0-h47c9636_0_cpython
  python-dateutil-2.8.2-pyhd8ed1ab_0
  python-tzdata-2023.3-pyhd8ed1ab_0
  python_abi-3.12-4_cp312
  pytz-2023.3.post1-pyhd8ed1ab_0
  pyyaml-6.0.1-py312h02f2b3b_1
  readline-8.2-h92ec313_1
  scipy-1.11.3-py312h44883eb_1
  seaduck-0.1.3-pyhd8ed1ab_0
  setuptools-68.2.2-pyhd8ed1ab_0
  six-1.16.0-pyh6c4a22f_0
  tk-8.6.13-h5083fa2_1
  toolz-0.12.0-pyhd8ed1ab_0
  tzdata-2023c-h71feb2d_0
  wheel-0.41.3-pyhd8ed1ab_0
  xarray-2023.10.1-pyhd8ed1ab_0
  xz-5.2.6-h57fd34a_0
  yaml-0.2.5-h3422bc3_2
  zipp-3.17.0-pyhd8ed1ab_0

Proceed ([y]/n)? y

Preparing transaction: done
Verifying transaction: done
Executing transaction: done
(trop_dev) syed44@MacBook-Pro-6 ~ % conda deactivate
(base) syed44@MacBook-Pro-6 ~ % conda create -n seaduck-env
Collecting package metadata (current_repodata.json): done
Solving environment: done

==> WARNING: A newer version of conda exists. <==
  current version: 23.7.3
  latest version: 23.10.0

Please update conda by running

    $ conda update -n base -c conda-forge conda

Or to minimize the number of packages updated during conda update use

     conda install conda=23.10.0

## Package Plan ##

  environment location: /Users/syed44/miniconda3/envs/seaduck-env

Proceed ([y]/n)? y

Preparing transaction: done
Verifying transaction: done
Executing transaction: done
#
# To activate this environment, use
#
#     $ conda activate seaduck-env
#
# To deactivate an active environment, use
#
#     $ conda deactivate

(base) syed44@MacBook-Pro-6 ~ % conda activate seaduck-env
(seaduck-env) syed44@MacBook-Pro-6 ~ % mamba install seaduck

Looking for: ['seaduck']

conda-forge/osx-arm64                                       Using cache
conda-forge/noarch                                          Using cache
pkgs/main/noarch                                              No change
pkgs/r/osx-arm64                                              No change
pkgs/r/noarch                                                 No change
pkgs/main/osx-arm64                                           No change
Transaction

  Prefix: /Users/syed44/miniconda3/envs/seaduck-env

  Updating specs:

   - seaduck

  Package                    Version  Build                 Channel           Size
────────────────────────────────────────────────────────────────────────────────────
  Install:
────────────────────────────────────────────────────────────────────────────────────

  + libexpat                   2.5.0  hb7217d7_1            conda-forge     Cached
  + python_abi                  3.12  4_cp312               conda-forge     Cached
  + libcxx                    16.0.6  h4653b0c_0            conda-forge     Cached
  + yaml                       0.2.5  h3422bc3_2            conda-forge     Cached
  + bzip2                      1.0.8  h93a5062_5            conda-forge     Cached
  + libffi                     3.4.2  h3422bc3_5            conda-forge     Cached
  + libzlib                   1.2.13  h53f4e23_5            conda-forge     Cached
  + xz                         5.2.6  h57fd34a_0            conda-forge     Cached
  + ncurses                      6.4  h463b476_2            conda-forge     Cached
  + ca-certificates        2023.7.22  hf0a4a13_0            conda-forge     Cached
  + llvm-openmp               17.0.4  hcd81f8e_0            conda-forge     Cached
  + tk                        8.6.13  h5083fa2_1            conda-forge     Cached
  + libsqlite                 3.44.0  h091b4b1_0            conda-forge     Cached
  + readline                     8.2  h92ec313_1            conda-forge     Cached
  + openssl                    3.1.4  h0d3ecfb_0            conda-forge     Cached
  + libgfortran5              13.2.0  hf226fd6_1            conda-forge     Cached
  + libgfortran                5.0.0  13_2_0_hd922786_1     conda-forge     Cached
  + libopenblas               0.3.24  openmp_hd76b1f2_0     conda-forge     Cached
  + libblas                    3.9.0  19_osxarm64_openblas  conda-forge     Cached
  + libcblas                   3.9.0  19_osxarm64_openblas  conda-forge     Cached
  + liblapack                  3.9.0  19_osxarm64_openblas  conda-forge     Cached
  + tzdata                     2023c  h71feb2d_0            conda-forge     Cached
  + python                    3.12.0  h47c9636_0_cpython    conda-forge     Cached
  + wheel                     0.41.3  pyhd8ed1ab_0          conda-forge     Cached
  + setuptools                68.2.2  pyhd8ed1ab_0          conda-forge     Cached
  + pip                       23.3.1  pyhd8ed1ab_0          conda-forge     Cached
  + six                       1.16.0  pyh6c4a22f_0          conda-forge     Cached
  + pytz                2023.3.post1  pyhd8ed1ab_0          conda-forge     Cached
  + python-tzdata             2023.3  pyhd8ed1ab_0          conda-forge     Cached
  + zipp                      3.17.0  pyhd8ed1ab_0          conda-forge     Cached
  + locket                     1.0.0  pyhd8ed1ab_0          conda-forge     Cached
  + toolz                     0.12.0  pyhd8ed1ab_0          conda-forge     Cached
  + packaging                   23.2  pyhd8ed1ab_0          conda-forge     Cached
  + fsspec                 2023.10.0  pyhca7485f_0          conda-forge     Cached
  + cloudpickle                3.0.0  pyhd8ed1ab_0          conda-forge     Cached
  + click                      8.1.7  unix_pyh707e725_0     conda-forge     Cached
  + python-dateutil            2.8.2  pyhd8ed1ab_0          conda-forge     Cached
  + importlib-metadata         6.8.0  pyha770c72_0          conda-forge     Cached
  + partd                      1.4.1  pyhd8ed1ab_0          conda-forge     Cached
  + importlib_metadata         6.8.0  hd8ed1ab_0            conda-forge     Cached
  + numpy                     1.26.0  py312h696b312_0       conda-forge     Cached
  + pyyaml                     6.0.1  py312h02f2b3b_1       conda-forge     Cached
  + pandas                     2.1.2  py312h9e53831_0       conda-forge     Cached
  + scipy                     1.11.3  py312h44883eb_1       conda-forge     Cached
  + dask-core              2023.10.1  pyhd8ed1ab_0          conda-forge     Cached
  + xarray                 2023.10.1  pyhd8ed1ab_0          conda-forge     Cached
  + seaduck                    0.1.3  pyhd8ed1ab_0          conda-forge     Cached

  Summary:

  Install: 47 packages

  Total download: 0 B

────────────────────────────────────────────────────────────────────────────────────

Confirm changes: [Y/n] y

Downloading and Extracting Packages

Preparing transaction: done
Verifying transaction: done
Executing transaction: done
```
