### 2023-11-08 MaceKuailv/seaduck#102 @syedhamidali
I would also like to mention that `mamba install seaduck` is working, but I haven't tested the package yet.
