### 2023-11-08 MaceKuailv/seaduck#102 @MaceKuailv
This is a little weird. It sure seems like all my machines are working fine. I will continue exploring this issue. But in order for you to continue looking at other parts of the package you can try several things.
1. try using mamba: you can `conda install mamba' and then replace every "conda" in the command with "mamba". Conda is a bit large and heavy, and sometimes there are weird issues that no one can figure out.
2. try pip installing it
3. install the development version following this guide (https://macekuailv.github.io/seaduck/contribute/prep_env.html). `pip install .` is pretty much guaranteed to work.
