### 2023-11-08 MaceKuailv/seaduck#102 @MaceKuailv
BTW, I have been looking at similar issues online and found several possible solutions. Those solutions are all trying to deal with the conda environment on the user's side rather than adding the package to conda-forge.
1. "re-acquaint" with conda-forge
```
conda config --append channels conda-forge
```
2. Totally refresh conda
```
conda clean -all
conda clean --index-cache
conda update conda
```
3. If you are offline (not likely), this should solve it
```
conda config --set offline false
```
Anyway, the fact that my machines (and also my friend's laptop with her newly installed conda) and your mamba can find the packages seems to suggest that the package is properly registered with conda-forge.

We are going to have a new release shortly as mentioned in issue #94, we can also try again then.
