### 2023-11-20 #5967 @MaceKuailv
Yes! I believe I am able to do that. The remaining issues from @pnsaevik will be resolved when I finish this round of revision and merge the PR. There is a good chance that the conda install issue raised by @syedhamidali will also be resolved when we have a new release.
