### 2023-11-20 MaceKuailv/seaduck#105 @syedhamidali
Clarification about the content of the paper.md
Are you sure about the word "shear" volume in the paper's opening paragraph?
Consider revising the word 'shear' to 'sheer' in this context. The term 'sheer' is more appropriate when referring to the vast or complete extent of something, like the volume of the Earth's oceans. 'Shear' typically relates to a force causing parts of a material to slide past one another, which doesn't seem to align with the intended meaning here.

This issue is a part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/5967
