### 2023-11-20 MaceKuailv/seaduck#106 @MaceKuailv
Good catch.

The methods you mentioned except for show_alias are considered internal methods that are not intended for the users. Pål has also pointed this out, but I didn't make the separation carefully enough. This will be solved in the next commit.
