### 2023-11-20 dd62c84
add docstring to some method and hide others
