### 2023-11-20 MaceKuailv/seaduck#105 @syedhamidali
@MaceKuailv, I recommend ensuring consistency in the language style used throughout the text. For instance, if you're adopting American English, certain words will need to be updated to reflect this choice - for example, changing 'behaviour' to 'behavior'. Maintaining this consistency is important for the coherence and professionalism of the document.
