### 2023-11-20 MaceKuailv/seaduck#105 @MaceKuailv
Thanks. I have done some additional proofread in the PR.

I apologize on behalf of my advisor @ThomasHaine for being British. :smile:
