+++ .pre-commit-config.yaml
@@ -16,7 +16,7 @@ repos:
   - id: nbstripout
     args: [--drop-empty-cells, --extra-keys "metadata.kernelspec"]
 - repo: https://github.com/psf/black
-  rev: 23.10.1
+  rev: 23.11.0
   hooks:
   - id: black
 - repo: https://github.com/keewis/blackdoc
@@ -25,7 +25,7 @@ repos:
   - id: blackdoc
     additional_dependencies: [black==22.3.0]
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.1.4
+  rev: v0.1.6
   hooks:
   - id: ruff
     args: [--fix]
