### 2023-11-21 MaceKuailv/seaduck#102 @MaceKuailv
@syedhamidali I have made a new release of the package. I have confirmed that I could install the new release 1.0.0 from two different machines. Can you try it from your side and let me know if it works?

If you are able to install, feel free to close this issue.
