### 2023-11-21 MaceKuailv/seaduck#101 @MaceKuailv
I have uploaded it here (https://zenodo.org/records/10183124).

Does it looks good to you?
