### 2023-11-21 #5967 @MaceKuailv
@kthyng I have uploaded the latest release to zenodo at (https://zenodo.org/records/10183124). The version is v1.0.0, and the doi is https://zenodo.org/doi/10.5281/zenodo.10183123

The only thing left on my table is the conda install issue https://github.com/MaceKuailv/seaduck/issues/102. I am waiting for @syedhamidali to get back to me after the latest release.
