### 2023-11-21 MaceKuailv/seaduck#101 @ThomasHaine
looks good to me except there's a typo in the title

Sent via the Samsung Galaxy S21 Ultra 5G, an AT&T 5G smartphone
Get Outlook for Android<https://aka.ms/AAb9ysg>
