### 2023-11-22 MaceKuailv/seaduck#102 @syedhamidali
It works now, although I had to reinstall Miniconda on my M1 machine.
