### 2023-11-22 #5967 @syedhamidali
@MaceKuailv, I've marked issue https://github.com/MaceKuailv/seaduck/issues/102 as closed. @kthyng, it appears that my portion of the task is complete. If there are any additional tasks or requirements, kindly inform me.
