### 2023-11-22 MaceKuailv/seaduck#102 @MaceKuailv
I had to do similar things on my machine recently as well. It is just very annoying.

I am glad it is working now.
