### 2023-11-22 d1e72d3
fix links and renew scisever after 1.0.0
