+++ docs/sciserver_notebooks/LLC4320.md
@@ -14,7 +14,7 @@ kernelspec:
 # Plot surface streamlines on LLC4320
 
 Author: Wenrui Jiang 14 June, 2023
-> **Warning**⚠️ : the notebook was last ran on **2023-07-27** with **seaduck 0.1.3**. You can find the executable version at https://github.com/MaceKuailv/seaduck_sciserver_notebook/blob/master/LLC4320.ipynb.
+> **Warning**⚠️ : the notebook was last ran on **2023-11-22** with **seaduck 1.0.0**. You can find the executable version at https://github.com/MaceKuailv/seaduck_sciserver_notebook/blob/master/LLC4320.ipynb. 
 The LLC4320 ocean circulation model solution is a kilometer-scale, global simulation with complex grid topology. This is a good dataset to test the performance of the `seaduck` package.
 
 ```{code-cell} ipython3
@@ -33,7 +33,7 @@ mpl.rcParams["figure.dpi"] = 600
 
 ```{admonition} Access LLC4320
 The global MITgcm run is the LLC4320 simulation [(Rocha
-et al. 2016)](https://journals.ametsoc.org/view/journals/phoc/46/2/jpo-d-15-0087.1.xmlhttps://journals.ametsoc.org/view/journals/phoc/46/2/jpo-d-15-0087.1.xml) and is publicly available on [SciServer](https://sciserver.org) (from the Oceanography container). The simulation output can be opened using the [OceanSpy](https://github.com/hainegroup/oceanspy) package using the [`from_catalog`](https://oceanspy.readthedocs.io/en/latest/generated/oceanspy.open_oceandataset.from_catalog.html#oceanspy.open_oceandataset.from_catalog) method.
+et al. 2016)](https://journals.ametsoc.org/view/journals/phoc/46/2/jpo-d-15-0087.1.xml) and is publicly available on [SciServer](https://sciserver.org/) (from the Oceanography container). The simulation output can be opened using the [OceanSpy](https://github.com/hainegroup/oceanspy) package using the [`from_catalog`](https://oceanspy.readthedocs.io/en/latest/generated/oceanspy.open_oceandataset.from_catalog.html) method. 
 
 `od = ospy.open_oceandataset.from_catalog("LLC4320")`
 
@@ -52,7 +52,7 @@ ds = od._ds
 oce = sd.OceData(ds)
 ```
 
-Initiate the particles randomly, so that they're distributed evenly on the globe. Use `N` = $1.5 \\times 10^5$ particles.
+Initiate the particles randomly, so that they're distributed evenly on the globe. Use $N = 1.5 \\times 10^5$ particles.
 
 ```{code-cell} ipython3
 :tags: [hide-input]
@@ -75,7 +75,7 @@ ax.plot(x, y, "o", markersize=0.1, transform=ccrs.PlateCarree())
 ax.coastlines()
 plt.show()
 ```
-![png](https://github.com/MaceKuailv/seaduck_sciserver_notebook/blob/master/LLC4320_files/LLC4320_5_0.png?raw=true)
+![png](https://github.com/MaceKuailv/seaduck_sciserver_notebook/blob/master/LLC4320_files/LLC4320_6_0.png?raw=true)
 
 We'll use just one vertical layer (the first one, which is at the surface) of a single snapshot (the first one) to make the streamlines.
 
@@ -250,7 +250,7 @@ pretty_stream_plot(
 )
 plt.show()
 ```
-![png](https://github.com/MaceKuailv/seaduck_sciserver_notebook/blob/master/LLC4320_files/LLC4320_29_2.png?raw=true)
[…]
