### 2023-11-22 69e485c
fix links and renew scisever after 1.0.0
