+++ docs/get_started.md
@@ -29,6 +29,7 @@ Running Lagrangian particle on regional dataset
 
 Example with complex grid (LLC4320)
 ^^^
+
 ```{image} https://github.com/MaceKuailv/seaduck_sciserver_notebook/blob/master/stable_images/LLC4320Thumbnail.png?raw=true
 :height: 200
 ```
