### 2023-11-22 MaceKuailv/seaduck#97 @MaceKuailv
This is resolved by PR 107
