### 2023-11-27 #5967 @kthyng
@MaceKuailv I have a few comments on your paper:
* A general comment is that despite being in the same field as you and doing similar work, I've had a hard time understanding your descriptions of what this tool does, until the paragraph before your "usage examples", and then in the usage examples themselves. Why not lead with the description of the tool as being for interpolation (Eulerian implied) and Lagrangian tracking (interpolation implied), or at least bring that idea out earlier in the paper. The descriptions earlier in the paper to me are more complicated and I've never heard of "lagrangian interpolation" as a concept (though it does give the tool a nice symmetry in the name).
* Does this tool address vertical interpolation? For example, xgcm can be used to interpolate from parametric vertical coordinates to depth coordinates.
* The references don't preserve capitalization currently but you can do so by adding {} around strings
* There are better references for TRACMASS than a conference abstract!
* Please change your doi's to not include the "http" etc — see the example paper for details: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
