### 2023-11-27 #5967 @kthyng
Thinking of more questions:
* Please expand the comparison in the paper to include other interpolation packages, for example xESMF, maybe scipy interpolation.
* xESMF accurately interpolates in lon/lat without projecting the coordinates — how does this work in seaduck and it's worth addressing that briefly in the paper.
