### 2023-11-27 #5967 @MaceKuailv
Thanks!

1. Sorry for the confusion. I will revisit later today. I agree Lagrangian interpolation is very rare in oceanography literature.
2. Yes, as long as the actual depth level is known.
3. Can you give me an example of that?
4. I will fix the TRACMASS reference.
5. I will fix the doi
6. &  7. I thought about talking about comparing with xESMF in the paper, but eventually didn't. For Eulerian interpolation, the result is going to be similar (since both are correct in my perspective). We offer some extra flexibility in the interpolation scheme, but the details are a little technical. For example, you can have land-aware adjustable interpolation kernel (https://macekuailv.github.io/seaduck/sciserver_notebooks/KangerFjord.html). However, our package is less flexible in the sense that we cannot do a global interpolation. For example, with scipy you could interpolate the entire field with a spline style interpolation.Yet, we would argue those sort of interpolation is not very practical on modern numerical models anyway. Personally, I also believe we have a simpler interface for those interpolations.

I am currently looking at a Lagrangian tracer budget constructed using this package, which is something no package that I know of can produce, so I am pretty confident that we are not reinventing wheels. I will revisit the text and make sure that this get across to the readers.
