### 2023-11-27 #5967 @kthyng
For capitalization:
Your oceanspy reference has for title `title = {OceanSpy: A Python package to facilitate ocean model data analysis and visualization},` but comes through with "python" lowercase. You can force it to stay capitalized with something like:
`title = {OceanSpy: A {Python} package to facilitate ocean model data analysis and visualization},` (or you can put the braces around only the "p" like {P}).

Regarding the other interpolation tools:
You don't need to argue that your tool is unique (nor add functionality to the package at this point to that effect), but you should add at least a paragraph in the paper comparing with the other relevant tools at a high level — the Statement of Need is a good place for this. What would someone who is interested in this tool want to know about this tool compared to others? Why use one vs. the other, etc, or what makes yours better or different? Something like the paragraph you just wrote in your comment is a good level, though there may be another interpolation tool or two to include. Probably there should be a paragraph on Lagrangian tracking models too.
