### 2023-11-27 #5967 @kthyng
I just remembered I forgot to say that your figure captions have "fig" in them which is redundant.
