+++ paper/paper.bib
@@ -30,12 +30,12 @@
    year = {2019},
 }
 @article{tracmass,
-   author = {Aitor Aldama Campino and Kristofer Döös and Sara Berglund and Dipanjan Dey and Joakim Kjellsson and Bror Jonsson and Aitor Aldama Campino and Kristofer Döös and Sara Berglund and Dipanjan Dey and Joakim Kjellsson and Bror Jonsson},
-   doi = {10.5194/EGUSPHERE-EGU2020-6984},
-   journal = {EGUGA},
-   pages = {6984},
-   title = {TRACMASS - A mass conserving trajectory code for ocean and atmosphere general circulation models},
-   url = {https://ui.adsabs.harvard.edu/abs/2020EGUGA..22.6984A/abstract},
+   author = {Aitor Aldama-Campino and Kristofer Döös and Joakim Kjellsson and Bror Jönsson},
+   doi = {10.5281/zenodo.4337926},
+   journal = {Zenodo},
+   keywords = {Lagrangian particle tracking,Numerical modelling,Offline},
+   title = {TRACMASS: Formal release of version 7.0},
+   url = {https://zenodo.org/records/4337926},
    year = {2020},
 }
 @article{Forget2015,
