### 2023-11-28 b8b49e2
bib capitalize, fig remove
