### 2023-11-28 a87d5b8
bib capitalize, fig remove
