+++ paper/paper.bib
@@ -24,7 +24,7 @@
    month = {7},
    pages = {1506},
    publisher = {The Open Journal},
-   title = {OceanSpy: A Python package to facilitate ocean model data analysis and visualization},
+   title = {OceanSpy: A {Python} package to facilitate ocean model data analysis and visualization},
    volume = {4},
    url = {https://joss.theoj.org/papers/10.21105/joss.01506},
    year = {2019},
@@ -62,7 +62,7 @@
    month = {8},
    pages = {3571-3584},
    publisher = {Copernicus GmbH},
-   title = {The Parcels v2.0 Lagrangian framework: New field interpolation schemes},
+   title = {The {Parcels} v2.0 {Lagrangian} framework: New field interpolation schemes},
    volume = {12},
    year = {2019},
 }
@@ -77,7 +77,7 @@
    month = {4},
    pages = {10},
    publisher = {Ubiquity Press},
-   title = {xarray: N-D labeled Arrays and Datasets in Python},
+   title = {xarray: N-D labeled {Arrays} and {Datasets} in {Python}},
    volume = {5},
    url = {https://openresearchsoftware.metajnl.com/articles/10.5334/jors.148},
    year = {2017},
@@ -94,7 +94,7 @@
    pages = {261-272},
    pmid = {32015543},
    publisher = {Nature Publishing Group},
-   title = {SciPy 1.0: fundamental algorithms for scientific computing in Python},
+   title = {SciPy 1.0: fundamental algorithms for scientific computing in {Python}},
    volume = {17},
    url = {https://www.nature.com/articles/s41592-019-0686-2},
    year = {2020},
@@ -108,7 +108,7 @@
    keywords = {Compiler,LLVM,Performance Keywords LLVM,Python},
    month = {11},
    publisher = {Association for Computing Machinery},
-   title = {Numba: A LLVM-based Python JIT Compiler},
+   title = {Numba: A {L}{L}{V}{M}-based {Python} {J}{I}{T} {Compiler}},
    volume = {2015-January},
    url = {https://dl.acm.org/doi/10.1145/2833157.2833162},
    year = {2015},
@@ -124,7 +124,7 @@
    month = {2},
    pages = {601-620},
    publisher = {American Meteorological Society},
-   title = {Mesoscale to Submesoscale Wavenumber Spectra in Drake Passage},
+   title = {Mesoscale to {Submesoscale} {Wavenumber} {Spectra} in {Drake} {Passage}},
    volume = {46},
    url = {https://journals.ametsoc.org/view/journals/phoc/46/2/jpo-d-15-0087.1.xml},
    year = {2016},
@@ -140,7 +140,7 @@
    month = {12},
    pages = {9156-9177},
    publisher = {John Wiley & Sons, Ltd},
-   title = {Wintertime Fjord-Shelf Interaction and Ice Sheet Melting in Southeast Greenland},
+   title = {Wintertime {Fjord}-{Shelf} {Interaction} and {Ice} {Sheet} {Melting} in {Southeast} {Greenland}},
    volume = {123},
    url = {https://onlinelibrary.wiley.com/doi/full/10.1029/2018JC014435 https://onlinelibrary.wiley.com/doi/abs/10.1029/2018JC014435 https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2018JC014435},
    year = {2018},
@@ -168,7 +168,7 @@ astronomy.},
    keywords = {Computing,Data science,Databases,Education,Science platform,Server-side analytics},
    month = {1},
    publisher = {Elsevier B.V.},
-   title = {SciServer: a Science Platform for Astronomy and Beyond},
+   title = {SciServer: a {Science} {Platform} for {Astronomy} and {Beyond}},
    volume = {33},
    url = {https://arxiv.org/abs/2001.08619v2},
[…]
