### 2023-11-28 #5967 @MaceKuailv
@kthyng I am having a little bit of problem changing the bibliography format to remove the http heading. The DOIs in my paper.bib file already does not have http in it, but it was added by the github action. Normally, with latex the solution is to set `\bibliographystyle{stylename}`, but I don't know where is this option when I am using markdown file.
