+++ paper/paper.bib
@@ -29,15 +29,6 @@
    url = {https://joss.theoj.org/papers/10.21105/joss.01506},
    year = {2019},
 }
-@article{tracmass,
-   author = {Aitor Aldama-Campino and Kristofer Döös and Joakim Kjellsson and Bror Jönsson},
-   doi = {10.5281/zenodo.4337926},
-   journal = {Zenodo},
-   keywords = {Lagrangian particle tracking,Numerical modelling,Offline},
-   title = {TRACMASS: Formal release of version 7.0},
-   url = {https://zenodo.org/records/4337926},
-   year = {2020},
-}
 @article{Forget2015,
    abstract = {This paper presents the ECCO v4 non-linear inverse modeling framework and its baseline solution for the evolving ocean state over the period 1992-2011. Both components are publicly available and subjected to regular, automated regression tests. The modeling framework includes sets of global conformal grids, a global model setup, implementations of data constraints and control parameters, an interface to algorithmic differentiation, as well as a grid-independent, fully capable Matlab toolbox. The baseline ECCO v4 solution is a dynamically consistent ocean state estimate without unidentified sources of heat and buoyancy, which any interested user will be able to reproduce accurately. The solution is an acceptable fit to most data and has been found to be physically plausible in many respects, as documented here and in related publications. Users are being provided with capabilities to assess model-data misfits for themselves. The synergy between modeling and data synthesis is asserted through the joint presentation of the modeling framework and the state estimate. In particular, the inverse estimate of parameterized physics was instrumental in improving the fit to the observed hydrography, and becomes an integral part of the ocean model setup available for general use. More generally, a first assessment of the relative importance of external, parametric and structural model errors is presented. Parametric and external model uncertainties appear to be of comparable importance and dominate over structural model uncertainty. The results generally underline the importance of including turbulent transport parameters in the inverse problem.},
    author = {G. Forget and J. M. Campin and P. Heimbach and C. N. Hill and R. M. Ponte and C. Wunsch},
@@ -199,4 +190,49 @@ astronomy.},
    title = {A finite-volume, incompressible {Navier} {Stokes} model for studies of the ocean on parallel computers},
    volume = {102},
    year = {1997},
+}
+@software{xesmf,
+  author       = {Jiawei Zhuang and
+                  raphael dussin and
+                  David Huard and
+                  Pascal Bourgault and
+                  Anderson Banihirwe and
+                  Stephane Raynaud and
+                  Brewster Malevich and
+                  Martin Schupfner and
+                  Filipe and
+                  Sam Levang and
+                  Charles Gauthier and
+                  André Jüling and
+                  Mattia Almansi and
+                  RichardScottOZ and
+                  RondeauG and
+                  Stephan Rasp and
+                  Trevor James Smith and
+                  Jemma Stachelek and
+                  Matthew Plough and
+                  Pierre and
+                  Ray Bell and
+                  Romain Caneill and
+                  Xianxiang Li},
+  title        = {pangeo-data/xESMF: v0.8.2},
+  month        = sep,
+  year         = 2023,
+  publisher    = {Zenodo},
+  version      = {v0.8.2},
+  doi          = {10.5281/zenodo.8356796},
+  url          = {https://doi.org/10.5281/zenodo.8356796}
+}
+@software{tracmass,
+  author       = {Aldama-Campino, Aitor and
+                  Döös, Kristofer and
+                  Kjellsson, Joakim and
+                  Jönsson, Bror},
+  title        = {TRACMASS: Formal release of version 7.0},
+  month        = dec,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v7.0-beta},
+  doi          = {10.5281/zenodo.4337926},
+  url          = {https://doi.org/10.5281/zenodo.4337926}
 }
