+++ paper/paper.md
@@ -37,7 +37,7 @@ The seaduck package is different from other ocean analytical tools (e.g., oceans
 
 The points can be stationary (fixed in space, or Eulerian) or be advected by a vector velocity field (Lagrangian). Most Lagrangian particle packages (e.g., [@oceanparcel, @individualdisplacement]) compute particle trajectories by solving the initial value problem numerically. Instead, seaduck uses efficient, accurate, mass-conserving analytic formulae, which assumes a step-wise steady velocity field similar to that used by TRACMASS [@tracmass]. The Lagrangian advection code is largely numba compiled, and the total amount of computation is less than solving the problem numerically. The Lagrangian particle functionality is based on the above-mentioned interpolation utilities, thus, it automatically navigates the complex topology of numerical ocean models.
 
-Compared to other interpolation tools (e.g. scipy[@Virtanen2020] and xESMF[xesmf]), Seaduck only supports Lagrangian polynomial interpolations using neighboring grid points. However, users can control a lot of other properties of a hierarchy of kernels which is not easy to implement in other packages. This includes: (1) The shape of the interpolation kernel(s) in both spatial and temporal dimensions, which defines which neighboring points are used, and therefore how the continuous field is estimated. (2) The interpolation weight function, which allows users to calculate generic linear operations on the data, such as differentiation and smoothing, in all four dimensions. The hierarchy of kernels controls behavior near land-masked points. Specifically, the hierarchy consists of successively more compact kernels that are used depending on the proximity of land points.
+Seaduck's interpolation functionalities works similar to other interpolation tools (e.g. scipy[@Virtanen2020] and xESMF[xesmf]). Despite that seaduck only support interpolation using neighboring grid points, users has control over a lot of properties of a hierarchy of the interpolation kernels which is not always present or easy to implement in other packages. This includes: (1) The shape of the interpolation kernel(s) in both spatial and temporal dimensions, which defines which neighboring points are used, and therefore how the continuous field is estimated. (2) The interpolation weight function, which allows users to calculate generic linear operations on the data beyond simple interpolation, such as differentiation and smoothing.(3) The hierarchy of kernels which controls behavior near land-masked points. Specifically, the hierarchy consists of successively more compact kernels that are used depending on the proximity of land points.Those interpolation functionalities are available in all four dimensions and for both Eulerian and Lagrangian points.
 
 With the above functionality, seaduck can accomplish many common tasks in ocean model data analysis, including interpolation, regridding, and Lagrangian particle simulation. Less common tasks are also possible, such as interpolation in Lagrangian label space, and analysis of tracer budgets along Lagrangian trajectories. We also strive to make seaduck an accessible education tool by creating a very simple high-level default interface, which is intended for people with little programming background, and for people who want to quickly try the tool.
 
