+++ Makefile
@@ -25,4 +25,4 @@ link-check:
 	jupyter-book build docs/ -W --keep-going --builder linkcheck
 
 joss:
-	pandoc paper/paper.md --bibliography paper/paper.bib -o paper/paper_local.pdf
+	pandoc paper/paper.md --bibliography paper/paper.bib -o paper/paper_local.pdf 
