+++ paper/paper.md
@@ -42,7 +42,7 @@ The seaduck package is different from other ocean analytical tools (e.g., oceans
 
 The points can be stationary (fixed in space, or Eulerian) or be advected by a vector velocity field (Lagrangian). Most Lagrangian particle packages (e.g., [@oceanparcel, @individualdisplacement]) compute particle trajectories by solving the initial value problem numerically. Instead, seaduck uses efficient, accurate, mass-conserving analytic formulae, which assumes a step-wise steady velocity field similar to that used by TRACMASS [@tracmass]. The Lagrangian advection code is largely numba compiled, and the total amount of computation is less than solving the problem numerically. The Lagrangian particle functionality is based on the above-mentioned interpolation utilities, thus, it automatically navigates the complex topology of numerical ocean models.
 
-Seaduck's interpolation functionality is similar to other interpolation tools (e.g., scipy[@Virtanen2020] and xESMF[xesmf]). 
+Seaduck's interpolation functionality is similar to other interpolation tools (e.g., scipy[@Virtanen2020] and xESMF[xesmf]).
 Seaduck only supports interpolation using neighboring grid points, but users control the properties of a hierarchy of interpolation kernels; which is an advantage. These properties include: (1) The space-time shape of the interpolation kernel(s), which defines which neighboring points are used, and therefore how the continuous field is estimated. (2) The kernel weight function, which allows users to calculate generic linear operations on the data beyond simple interpolation, such as differentiation and smoothing. (3) The hierarchy of kernels near land-masked points.
 
 With the above functionality, seaduck can accomplish many common tasks in ocean model data analysis, including interpolation, regridding, and Lagrangian particle simulation. Less common tasks are also possible, such as interpolation in Lagrangian label space, and analysis of tracer budgets along Lagrangian trajectories. We also strive to make seaduck an accessible education tool by creating a very simple high-level default interface, which is intended for people with little programming background, and for people who want to quickly try the tool.
