### 2023-11-29 #5967 @MaceKuailv
@kthyng The latest PR has included changed to address the issues you mentioned, except for the one with http heading. The URL in our .bib file is clean, it does not contain the http heading. That is added in the github action. I have looked at a few other JOSS paper under review, they all have http for their doi (e.g. https://raw.githubusercontent.com/openjournals/joss-papers/joss.05982/joss.05982/10.21105.joss.05982.pdf).

I have tried a lot of things. Something that worked on my laptop is to add a metadata option "link-bibliography" (https://pandoc.org/MANUAL.html#other-relevant-metadata-fields). However, this still does not get rid of the http heading generated here. I could test it some more here, but it is going to be a lot of disturbance for all of you.

Please let me know what I should do now.
