### 2023-12-04 #5967 @kthyng
@MaceKuailv Ok it's looking pretty good though the references you added for scipy and xESMF aren't formatted properly in the paper — you should be able to match their syntax to the other references to have it look right. Also there is a gap in line 68.
