### 2023-12-04 3f6df2b
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/nbQA-dev/nbQA: 1.7.0 → 1.7.1](https://github.com/nbQA-dev/nbQA/compare/1.7.0...1.7.1)
