+++ .pre-commit-config.yaml
@@ -45,7 +45,7 @@ repos:
   - id: pretty-format-toml
     args: [--autofix]
 - repo: https://github.com/nbQA-dev/nbQA
-  rev: 1.7.0
+  rev: 1.7.1
   hooks:
   - id: nbqa-black
   - id: nbqa-ruff
