### 2023-12-19 #5967 @MaceKuailv
I have a very large image which could potentially cause a timeout. Is that what happened?
